adds optional pagination params to WebhookLog#list
It's odd that we include a "next_url" key (see below) and then offer
no way to request that URL, or its equivalent, through the FundAmerica gem
instead of using API::request directly.

Since page defaults to 1 and per defaults to 25, it looks like this gem lacks
any way to get details about WebhookLogs that are older than the 25 most recent.

(Perhaps this is a part of the gem that is still under construction, or no one
has needed to check more than 25 at a time, previously?)

Example output:

[1] pry(main)> FundAmerica::WebhookLog.list
=> {"object"=>"resource_list",
    "total_resources"=>40,
    "page"=>1,
    "per"=>25,
    "more"=>true,
    "next_url"=>"https://sandbox.fundamerica.com/api/webhook_logs?page=2&per=25",
    "resources"=>
      [{"object"=>"webhook_log",
      ...